requirements.txt
pytz
requests
python-dotenv

ipython==8.9.0  # enhanced python interpreter

# Django
django==4.1.6
django-cors-headers
django-debug-toolbar==3.8.1
whitenoise==6.3.0
django-dbsettings # for storing settings in database

# Django 3.0 async requirements
gunicorn==20.1.0
uvicorn==0.20.0

# Databases
psycopg2-binary==2.9.5
dj-database-url==1.2.0

# Distributed async tasks
celery==5.2.7
redis==4.4.2
django-celery-beat==2.4.0

# Telegram
python-telegram-bot==13.15  # last sync version

# Typing
pydantic

# Payments
cloudpayments

pillow
qrcode==7.4
